requirement.txt
python-dotenv   
aiogram
openai